refactor, cleanup, bugfixes
update gradle to 7.4
add copyright header
remove NPC contact code
update plugin descriptor
use dialogue for repair check
add hub icon
