+++ LICENSE
@@ -1,6 +1,7 @@
 BSD 2-Clause License
 
 Copyright (c) 2020, David Ventura
+Copyright (c) 2023, Truth Forger <http://github.com/Blackberry0Pie>
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
