+++ runelite-plugin.properties
@@ -1,6 +1,6 @@
-displayName=Pouch time to degrade
-author=DavidV
+displayName=Pouch Health
+author=DavidV,Truth Forger
 support=https://github.com/DavidVentura/rc-pouch-alert
-description=Shows how much essence you can put into your pouches before it degrades. Resets only when casting Contact.
-tags=rc,runecraft,runecrafting,pouch
-plugins=com.example.PouchUsageLeft
+description=Shows how much essence you can put into your pouches before it degrades.
+tags=rc,runecraft,runecrafting,pouch,rune,ess,degrade
+plugins=ventura.rcpouchalert.PouchUsageLeftPlugin
